pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "visionflow"
version = "0.1.0"
description = "Event-driven, real-time AI video stream processing framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [
    {name = "VisionFlow Contributors"},
]
keywords = ["video", "streaming", "AI", "YOLO", "OCR", "event-driven", "real-time"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "opencv-python>=4.8.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.0",
    "aiofiles>=23.2.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
yolo = ["ultralytics>=8.0.0"]
ocr = ["pytesseract>=0.3.10", "pillow>=10.0.0"]
kafka = ["kafka-python>=2.0.2"]
livekit = ["livekit>=0.8.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.scripts]
visionflow = "visionflow.cli.main:main"

[tool.setuptools]
packages = ["visionflow"]

[tool.setuptools.package-data]
visionflow = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
